Dev feature pi12 103 (#24)
* Scheduler

* Scheduler_time

* Update SchedulerTime

* Aggiunto dataset e cancellato delete dataset

* Scheduler Programma caricameno CSV funzionante e primo tentativo caricamento CSV Storico di Generation (no funzionante)

* Terna_Generation_Historical

* Program di caricamento dati da CSV a DB funzionante + Scheduler funzionante (attenzione allo Shutdown)

- CSV caricano Macrozone, Region, Province, Industry e lo storico (da 2021) di Generation (da finire storico di Load)
- ATTENZIONE ai README di ProgettoIndustriale.Data e TestManagerCSV

* Aggiunta storico Load. Modifiche a Domain, Dto e Program (TestManagerCSV) affinchè il caricamento di dati su DB di Load funzioni

+++ ProgettoIndustriale.Type/Domain/Load.cs
@@ -14,11 +14,10 @@ namespace ProgettoIndustriale.Type.Domain
         public int Id { get; set; }
 
         [Column("total_load_MW")]
-        public int TotalLoadMW { get; set; }
-
+        public double TotalLoadMW { get; set; }
         [Column("forecast_total_load_MW")]
-        public float ForecastTotalLoadMw { get; set; }
-        
+        public double ForecastTotalLoadMw { get; set; }
+
         [Column("COD_date")]
         public int IdDate { get; set; }
